feat: useSearch hook and refactor on ingredientSearch component
